Add CAN polling task and improve UI update locking mechanism
- Introduced a dedicated task for CAN polling and input engine updates to ensure responsiveness even when the main loop is busy.
- Added a new function `lvgl_port_lock_with_timeout()` to prevent indefinite blocking during UI updates.
- Updated `handle_can_messages()` to utilize the new locking mechanism with a timeout.
- Refactored the input engine loop to use an integer for the input check to ensure proper termination.
- Minor adjustments to the main loop to accommodate the new task structure and improve overall system stability.

+++ src/common/input_engine.cpp
@@ -145,8 +145,8 @@ void input_engine_process_level(uint8_t input_id, bool active_now) {
 void input_engine_update(void) {
   unsigned long now = millis();
   
-  // Check all inputs for pending click timeouts
-  for (uint8_t i = 0; i < 256; i++) {
+  // Check all inputs for pending click timeouts (use int so loop terminates at 256)
+  for (int i = 0; i < 256; i++) {
     input_state_t* state = &g_state[i];
     if (state->click_pending && (now - state->pending_click_time) > (unsigned long)g_double_click_gap_ms) {
       // Find config to verify this input is valid
